BPSO-3520 Fix: Add retry to Producer._next_partition
Additionally fix tests which fail as a result of fix for BPSO-3520.
Due to the restructuring, when we don't want the _next_parttion() call
to fail, we need to set the return value for the
client.metadata_error_for_topic mock to be True, otherwise the error
check in _next_partition will receive a 'Mock' object (which evaluates
to True) and never succeed.
Where the failure was desired, at least initially, a clock was needed to
exercise the DelayedCall waiting mechanism.