Address issue where commit() callers could interfer
If a caller calls commit() when another commit is in progress, a
deferred is returned.  Prior to this fix the deferred returned was the
same one returned to the caller of the original commit. Due to the
nature of deferreds, this meant that any callers but the first could see
the result returned in the deferred affected by the first caller. With
this change a separate deferred is used to track each caller's result.