BPPF-416: Add 'shutdown' method to Afkak Consumer
Added a shutdown method to the Afkak consumer object to allow
processors to gracefully stop the consumer and have any processed
messages' offsets committed.  If the processor has returned a deferred
prior to shutdown being called, then the shutdown waits on the
returned deferred. However, if the shutdown is called prior to the
deferred being returned then the consumer is stopped immediately after
any commit operation (as needed) and any deferred returned by the
producer is cancelled.

Also added a performance test that had been kicking around in my sandbox.