Log error when reactor is blocked during request
When the reactor is blocked during a request for longer than the timeout
for the request, the request will be cancelled and fail, even if the
Kafka broker has returned a result (a result of how Twisted processes
timer events before socket events).  This change will cause Afkak to log
an error message in this case, making debugging easier.