Fix 'CRITICAL twisted Unhandled error in Deferred:' log message
Afkak was failing to add an error handler to a deferred returned from
load_metadata_for_topics when it was called as part of the handling of
the drop of a connection to a Kafka broker. This resulted in a critical
log message from Twisted. This change adds the error handler and tests
the failure case to ensure it's properly handled in the client library,
rather than at the Twisted garbage collection level.

It includes improvements to handling of disconnected brokers:
Prefer connected brokers when sending broker unaware requests
KafkaBrokerClient now exposes a `connected` method to return whether
it is currently connected to a Kafka broker.
Removed unused `brokers` arg to
KafkaClient._send_broker_unaware_request.
Improve warning when requests fail by including the requestId

Also, it removes the optimization to re-use outstanding request to Kafka
for app-initiated calls to load_metadata_for_topics as it complicated
proper error handling.

Rework test of load_metadata_for_topics to ensure that queries are
treated independently.

In test, simplify failure returned from load_metadata_for_topics as it
was 'needlessly authentic'.

Update test for new logging level of message logged when metadata fetch
fails after disconnect.

Increased max_wait_time in fetch request in test_roundtrip_large_request
as sometimes I'd see the timemout result in the test failing to retrieve
the message.