FIX BPSO-45921: Cleanup close_dlist attribute when deferred fires.
When updating the active KafkaBrokerClients as a result of loading
metadata from the Kafka cluster, Afkak would track the close() of the
broker-clients which were no longer part of the cluster. However, once
the close completed, it would clear the tracking deferredList instance
variable, and would instead add it to head of the next deferredList
tracking closes. This would result in a nested deferredList which grew
without bound and would eventually cause a failure when a debugging log
message would try to render it recursively.

This change adds a callback/errback handler to the deferredList which
clears the client's close_dlist instance variable (releasing the
deferredList) if that deferredList is currently the client's active,
top-level close_dlist.
Also simplified the clearing of client._collect_hosts_d.

Improved the testing around both client.close() and
client._update_brokers.