Refactor close of broker clients into separate function
In addressing Tom's feedback about handling the close deferred when
closing the clients, I realized that I should cleanup the duplicate code
used in both `client.close` and `client._update_brokers`, moving it to a
separate function.
NOTE: the deferred returned from `client.close` now properly yields just
a `None` value when complete, not the results of the individual close
operations which is internal to the client.