Properly reset consumer group metadata
Previously, KafkaClient's reset_all_metadata failed to reset the
consumer group metadata (consumer_group_to_brokers), so that when a
broker disconnected and the client called reset_all_metadata to force an
update, it failed to reset the consumer group metadata.  This could
result in attempting to connect a down broker for consumer group
operations, and could lead to adding an obsolete broker (ip/port) back
into the client's brokers map.  This could lead to long (10 second)
delays for operations which randomly selected a brokerclient.

This change adds the reset of the consumer group metadata to the
reset_all_metadata routine, and improves testing to ensure the metadata
is set and cleared properly.

Signed-off-by: Robert Thille rthille@ciena.com