Fix BPPF-330: Error on Kafka broker disconnect.
This turned out to be due to the failure of Afkak to properly reset the
delay before reconnect attempts when a successful connection was
made. This meant that after a few dropped connections to a single Kafka
broker, the delay before a reconnect was attempted would increase to 15
seconds. This is larger than the 10 second default request timeout. So,
on a connection drop, the automatic request for new metadata would be
initiated immediately but would timeout before a new connection was
brought up to the broker.

The fix was just to move the `self.resetDelay()` call to
`buildProtocol`, which is called whenever the connection successfully
completes, not from `_connect` which is only called once to build the
connector object.

In fixing, I decided to lower the initial reconnect delay from 1 second
to 0.1 seconds, and also add an `initDelay` parameter to the constructor
of the KafkaBrokerClient so that could be overridden.