Fix BPPF-472: Fix tight-loop blocking reactor on stop()
In some cases, stopping the consumer from within the processor could
cause the KafkaConsumer to enter into a tight loop with no yield to the
reactor.

Fixed by clearing the `self._msg_block_d` before cancelling so
`_handle_fetch_response` would not attempt to `addCallback()` to it
after it already had a result, resulting in a recursive infinite loop.

Also check in `_retry_fetch`, check for stopping/stopped condition and
do not schedule a fetch in such case.