Fix raised KafkaUnavailableError to have correct list of brokers
When trying a list of brokers, the client would raise an error if
requests failed to every broker, but it would raise the error with the
client's current list of brokers, not the list it actually attempted
to send through.

Also, fix comment/docstring for Producer init method which was missing
items and had the wrong name for one item.