Address review comments and fix BPPF-3752
Removed (most of) the crazy 'mockback*' mocks attached to deferreds in
favor of failureResultOf/successResultOf. Kept in the few places where
I was waiting on the deferred to fire in a loop, not checking the
result.

Introduced `RestartError` and `RestopError` for when a caller attempts
to start a running, or stop a stopped, consumer.

Fix BPPF-3752 by checking for 'None' as last_processed_offset before
attempting commit request.

Updated copyright headers.