Forbid brokerclient use and creation after close()
Adding the invariant that a broker client must be removed from the
client map before being closed. This makes the code easier to reason
about because the client map only contains live clients.

Add an explicit check to prevent creation or use of broker clients once
KafkaClient.close() is called. The clients themselves are poisoned by
calling the close() method in _close_brokerclients(). Further the client
map is set to None to prevent reuse.