Rework broker client for cleaner cancellation
These changes are motivated by difficulties debugging integration test
failures on #37.

Cancellation: One Way To Do It
==============================

Remove the _KafkaBrokerClient.cancelRequest() method that could cancel
a pending request. This was redundant with the cancel() method of the
returned deferred. Having two ways to do the same thing led to confusing
flow control (particularly as the deferred's cancel() method can be
called implicitly by twisted.internet.defer.inlineCallbacks).

Track Cancelled correlationIDs
==============================

A request that was cancelled would be removed from the
_KafkaBrokerClient.requests map immediately. If a response was later
received the broker would log a warning and a bit of the response. This
created lots of preventable log noise.

Now a cancelled request is tracked until it gets a response or the
connection drops. The log message for an _unexpected_ response has been
upped to an error, since it almost certainly represents a bug.

Shutdown is not Cancellation
============================

Calling _KafkaBrokerClient.close() would cancel all pending requests
with an afkak.common.CancelledError (not to be confused with
twisted.internet.defer.CancelledError, of which it is a subclass).
Higher level logic in KafkaClient, Consumer, and Producer would treat
cancellation specially (by not logging) or not specially enough (by
doing retries when shutting down).

Now _KafkaBrokerClient.close() terminates outstanding requests with
afkak.common.ClientError, mirroring the behavior of
_KafkaBrokerClient.makeRequest() which rejects attempts with ClientError
once the broker client has been closed. This means that higher-level
code can treat ClientError as evidence of shutdown (which wasn't true of
CancelledError).

This change is not backwards compatible.

Other Changes
=============

- Be consistent about referring to correlation IDs as `correlationId`,
  rather than `requestId`, `id`, or `correlation_id`. This matches
  Kafka's own log messages and other parts of Afkak.
- Track the time each message is queued, sent, and cancelled. This
  information can assist in interactive debugging.