reset_topic_metadata() should always clear errors
When a topic is present in KafkaClient.topic_errors calling
reset_topic_metadata() didn't clear it unless it also existed in also
exist in KafkaClient.topic_partitons — i.e., it had some partitions.
This could prevent error state from being cleared, but was probably
mostly masked in practice by the aggressive resetting of metadata in
other parts of the client.