Rework test_consumer_rejoin assertions
This integration test was fixed by making the ConsumerGroup rejoin on
UnknownMemberId. These changes were made while debugging the failure.
Rather than using monkey-patching to watch for internal code paths which
makes the test brittle, these changes make it wait for the desired state
to appear. They also validate more about how partitions are assigned
than the previous assertions.

I had to stop the test from waiting on the deferred returned by
ConsumerGroup(...).start(). It doesn't seem to fire reliably
— I observed tests hanging forever in the cleanup phase waiting on this
deferred.

I am unsure what this test is trying to cover. It disables the heartbeat
of one group member (by poking at private internals), then adds a second
group member. Adding the second group member triggers a rebalance which
restores the first member to a valid state. Is this meant to imitate the
reactor of one of the members getting blocked for a while? Couldn't we
test this sort of scenario better with with a Jepsen-style test?