Signed-off-by: Paul Hudgins <phudgins@ciena.com>
The setting of OFFSET_LATEST for auto_offset_reset was not getting honored.  It was always setting it to OFFSET_EARLIEST when the consumer group was not found.  This change will check for a passed-in value to use in this case.  The default, when None is passed in, is to set it to OFFSET_EARLIEST.